+++ firmware/tinyg/stepper.h
@@ -392,8 +392,7 @@ stat_t st_motor_power_callback(void);
 void st_request_exec_move(void);
 void st_prep_null(void);
 void st_prep_dwell(double microseconds);
-//stat_t st_prep_line(float steps[], float microseconds, uint8_t last_segment_flagged);
-stat_t st_prep_line(float steps[], float microseconds, int32_t encoder_error[]);
+stat_t st_prep_line(float steps[], float microseconds, float encoder_error[]);
 
 stat_t st_set_sa(cmdObj_t *cmd);
 stat_t st_set_tr(cmdObj_t *cmd);
